El driver de dispositivos de bloques BIOS (DB.C) ha enpezado a funcionar con un thread para atender las peticiones de los procesos correspondientes a las llamadas al sistema read/write. Puede probarse con el programa de usuario PROGDB.C. Por ejemplo: SO1> progdb FDA 100 100000. Ese comando lleva a cabo la lectura de 100000 bytes del disquete (FDA) a partir del byte 100. La lectura se fracciona en llamadas sucesivas a read(3, dir, 512). El programa funciona bien con la multiprogramación. Incluso podemos arrancar simultaneanente varios procesos que ejecuten ese comando, ejecutandose en paralelo sin ningún problema. Falta por hacer:
  - pulir rutina de interfaz LSEEK.C (retorno de la posicion)
  - pulir programa PROGDB.C llamándolo LEERDB.C
  - hacer un comando (o modificar INFO.C) para que muestre los
    datos de los dispositivos de bloques detectados por el BIOS.
    Para eso se utiliza la llamada ioctl(df, 1, &d_bloque).
  - Basar SO1 en ese driver en cuanto al sistema de ficheros y
    la llamada al sistema exec.

+++ PRACT0/SO1/MANEJ_01.C
@@ -38,7 +38,7 @@
 #include <so1.h\procesos.h>    /* descProceso, descFihero, descRecurso ... */ 
 #include <so1.h\minifs.h>                                        /* rec_sf */
 
-//#include <so1.h\bios.h>                                  /* printStrBIOS */
+#include <so1pub.h\bios_0.h>                               /* printStrBIOS */
 
 int indiceTFAS ( const char far * nombre ) 
 {
@@ -232,7 +232,7 @@ void so1_manejador_01 ( void )                         /* ah = 1 ; int SO1 */
                     if ((SEEK_SET <= whence) && (whence <= SEEK_END)) 
 					{
                         posNueva = descRecurso[rindx].lseek(dfs, pos, whence) ;
-                        if (posNueva >= (long)0)
+                        if (posNueva >= 0L)
                             ptrDPActual->tfa[df].pos = posNueva ;
                     }
                 }
